# Proyecto_plaGA/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Detección de Plagas",
        "description": "Documentación de la API para el sistema de detección de plagas usando Laravel y YOLOv5",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
        "/api/notificaciones": {
            "get": {
                "tags": [
                    "Notificaciones"
                ],
                "summary": "Obtener las últimas 10 notificaciones del usuario autenticado",
                "operationId": "4a5488228077528a278bea003da9a3cc",
                "responses": {
                    "200": {
                        "description": "Lista de notificaciones",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "notificaciones": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "mensaje": {
                                                        "type": "string",
                                                        "example": "Tienes una nueva alerta"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-07-19T20:00:00Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-07-19T20:00:00Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Notificaciones"
                ],
                "summary": "Crear una nueva notificación y emitirla en tiempo real",
                "operationId": "d2dd16a0c0f439306ecfa7ec0312972f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "mensaje"
                                ],
                                "properties": {
                                    "mensaje": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Tienes una nueva alerta"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Notificación creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "notificacion": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "mensaje": {
                                                    "type": "string",
                                                    "example": "Tienes una nueva alerta"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-07-19T20:00:00Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-07-19T20:00:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Registrar un nuevo usuario",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "cedula",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Juan Pérez"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan@example.com"
                                    },
                                    "cedula": {
                                        "type": "string",
                                        "example": "1234567890"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario registrado correctamente y token JWT retornado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Iniciar sesión y obtener token JWT",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token JWT retornado si las credenciales son válidas"
                    },
                    "401": {
                        "description": "Credenciales inválidas"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cerrar sesión (JWT, frontend debe eliminar el token)",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Logout exitoso"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/deteccion": {
            "get": {
                "tags": [
                    "Detecciones"
                ],
                "summary": "Obtener las detecciones del usuario autenticado",
                "operationId": "9b6bf8b8bba8b347a05565bdc4f5594f",
                "responses": {
                    "200": {
                        "description": "Lista de detecciones"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Detecciones"
                ],
                "summary": "Registrar una detección de plaga",
                "description": "Registra una nueva detección de plaga asociada al usuario autenticado. Este endpoint requiere autenticación (JWT).",
                "operationId": "610223a6c636bdb93a3cef61d9e4ed9f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "plaga",
                                    "ubicacion",
                                    "hora"
                                ],
                                "properties": {
                                    "plaga": {
                                        "type": "string",
                                        "example": "Mosca blanca"
                                    },
                                    "ubicacion": {
                                        "type": "string",
                                        "example": "Invernadero 3, sector norte"
                                    },
                                    "hora": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-07-19T14:30:00Z"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Detección registrada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensaje": {
                                            "type": "string",
                                            "example": "Detectada y guardada"
                                        },
                                        "id": {
                                            "type": "integer",
                                            "example": 42
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/plagas/detectar": {
            "post": {
                "tags": [
                    "Plagas"
                ],
                "summary": "Detectar plagas en una imagen base64 usando YOLOv5",
                "operationId": "7db10e707ecf8fc0e91891743f955bda",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "imagen"
                                ],
                                "properties": {
                                    "imagen": {
                                        "type": "string",
                                        "format": "base64",
                                        "example": "data:image/jpeg;base64,/9j/4AAQSkZJRgABAQAAAQABAAD..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Resultados de detección en JSON"
                    },
                    "400": {
                        "description": "Formato de imagen inválido o error en la detección"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Notificaciones",
            "description": "Notificaciones"
        },
        {
            "name": "Autenticación",
            "description": "Autenticación"
        },
        {
            "name": "Detecciones",
            "description": "Detecciones"
        },
        {
            "name": "Plagas",
            "description": "Plagas"
        }
    ]
}